git backup: generate backup branch from current HEAD
Store current HEAD in a backup branch, i.e. before rebasing your current
branch. As it is intended to keep the current state of the branch as a
backup (hence the name), it does _not_ checkout the new generated
branch, also it does _not_ implicitely push the branch to somewhere else.

git backup is equivalent to calling

git branch <current>-backup-<date>-<index>

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>