git-fake-git-patch-gz: prepend fake git header to compressed bare patch
git-fake-git-patch-gz prepends a fake git commit header to a gz
compressed patch file. The file is prepended in-place.

The usecase for this very special command is using devtool for
extracting a linux kernel with the preempt-rt patch, which when
downloaded is only a raw patch without the git metadata prepended.

Please note that this tool is very crude and just prepends a dummy
header without checking anything in the patch file. Your mileage may
vary.

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>